+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2019-08-11
+
+### Added
+
+- `detailed` option to opt-in for personal/detailed tracking
+
+### Changed
+
+- Attributes (`.attributes()`) don't include personal data by default
+
 ## [3.1.3] - 2019-03-29
 
 ### Changed